Move file out of netlify functions directory
