+++ tsconfig.json
@@ -18,5 +18,13 @@
     "incremental": true,
     "allowImportingTsExtensions": true
   },
-  "include": ["src", "tests", "test", "app", "netlify/functions", "vite.config.mts"]
+  "include": [
+    "src",
+    "tests",
+    "test",
+    "app",
+    "netlify/functions",
+    "netlify/shared",
+    "vite.config.mts"
+  ]
 }
